Add manuals-jobs for training-guides
The training-guides have their own repo now, but their manuals are no
longer published to docs.openstack.org.

The current infra-config contains only a partial solution. There is a
project in projects.yaml, but no matching job in manuals-jobs.yaml.

The first iteration of this patch had two shortcomings:

1) It was based on code (or rather config file) reading and missed
   another problem that had to be fixed first: a left-over from
   openstack-manuals tried to copy static files from a non-existent
   directory (see Change-Id: I43f56523a7fe59d2d9ea92cabfcb4e926e751f6d).

2) It added a manuals-job, but didn't add the job to the project.

With 1) fixed, training-guides publishes successfully to docs-draft. In
order for the material to appear on doc.openstack.org as well, we need
to call upload-to-docs-site (macros.yaml), too.

Closes-Bug: 1329567
Change-Id: I6308ee2eca021fc57ddfd83c6ea232703c66fb13