Make the fallback sysadmins value the same
The fallback value for sysadmins was was admins in some places
and admin in others. However, re: jeblair:

The fallback value for sysadmins should be "[]", which is actually the
default value. Configuring the mail system to send root's mail to an
account called 'admin' which may or may not exist and if it does exist
may send mail to root is not a good idea.

The default of [] will deliver root's mail normally.

Change-Id: Iaeca86ef135b6a3210541618d48caf4058dc7966