Create user to catch email if $sysadmin == []
When no sysadmin's are defined ($sysadmin == []) for the exim module
then the /etc/aliases file isn't completed correctly. The alias
for root isn't created meaning the root user is the catch all.

This doesn't work in exim, as the root user is on the disallowed
list. Meaning if $sysadmin == [], then exim tries to send email
to a disallowed user causing error messages to appear in the
exim error log.

To solve this, this change checks to see if $sysadmin == [], and
if so, creates a sysadmin user to catch the email. If $sysadmin
has been defined, then puppet makes sure this sysadmin users doesn't
exist.

Change-Id: I2f93d92f1bbb6692fdc919ba5738f21b4cad2340