demote f20 to experimental
as we're currently at quota, with about 1.5x quota worth of load
in the check queue. That's put us in a state that f20 nodes are
not being allocated at all. Which means that devstack and devstack-gate
changes are currently stalled.

The fix for the top gate bug is probably a devstack fix. Which puts us
in a position where we can't make things better because we're
stuck waiting for f20 nodes.

Move to experimental for now. We need a longer term fix to prevent
us from getting into this situation before moving it back to check.

Change-Id: I7f44cb9896f646060d45ac571ddb000b721812a6