Whitelist hacking in the requirements job
Due to the nature of hacking, changes to hacking in global-requirements
are often backwards incompatible since they introduce new rules. Because
of this we want to let projects adopt the new version of hacking at
there own pace.

An additional change will be made to the proposal bot to stop automatic
roll-out of hacking version changes from global-requirements.

Change-Id: I114b05fb1ec99ec1e0e0162988edb5fa5517df46