Adds Third Party CI account name format directions
We have new name formating for username and Full Name for Third
Part CI accounts for Gerrit. This patch adds the documentation
for this new formatting to the third_party.rst file.

This patch also changes the documentation about what methods can
be used to request an account. Since some developers were sending
emails AND filing bugs AND requesting them in channel, I removed
the options and identified sending an email as the sole source
of requesting an account.

Change-Id: Ia3f3a893bcdb67f81ebdf05506e2880c0c5b1a5c