Adds Third Party CI account name format directions
We have new name formatting for username and Full Name for Third
Party CI accounts for Gerrit. This patch adds the documentation
for this new formatting to the third_party.rst file.

Earlier versions of this patch included a change in what method
could be used to request a third party ci gerrit account, that
change has moved to: https://review.openstack.org/#/c/102538/

Change-Id: Ia3f3a893bcdb67f81ebdf05506e2880c0c5b1a5c