Sahara: add job to build swift fs for hadoop
We have a bunch of Java code in sahara-extra repo and we'd like to have
this code published on the tarballs.o.o to make our lifes easier. We're
working on moving this code to the Hadoop community (it's already
committed to one of the Hadoop versions but not GA yet).

Change-Id: I7760dd55dbfa7b1ebf58ed7e63c2aee261716499