Explicitly set node label for every zuul job.
In order to facilitate the transition from precise to trusty for Juno
testing we need to make zuul the decider of node labels for every job.
Do this by refactoring how we apply parameter functions to zuul jobs.

Since we can have only one parameter function per job it is easier to do
decision logic in a single global parameter function. Add this single
function and have it maintain backward compatbility with old parameter
function selections. But also make it select the ZUUL_NODE for every
job.

Change-Id: I42fa26fc42c7dc4e94d3a7382cc8de0a5e6d1f36