Fix cacti node and add os check
Previously a file was being managed that couldn't exist becuase
the parent directory did not exist. Cacti succeded in spite of
itself because the cacti package pushed an apache config into a
place that works.

This change sets the file path of the apache.conf file for cacti
to be a real location, which adds the rewrite rules. Previously,
hitting '/' would not redirect to the graph viewing page, despite
that being the intention of the puppet code.

This change also adds some logic to fail early on non Debian
machines, since we're explicitly using the /etc/apache2 directory
here.

Change-Id: I29c04d939d7a8d1a0faeb4ab7b31e3adc951ccf3