Add keystonemiddleware project
The OpenStack Identity (Keystone) team has decided to split the
packaging of the middleware used by external services (both from
Keystone and python-keystoneclient) into it's own separate packaging.

This is to enable the reduction in dependencies required by the
keystoneclient library (some dependencies are specific to servers
such as memcache) and allow for releasing fixes to the various
middleware code without needing to also release either a fix to
Keystone or a fix to the keystoneclient library.

The new keystonemiddleware project will be released independent
of the OpenStack named-cycle (similar to the client libraries).

Further details on this change can be seen in the approved
Keystone specification: https://review.openstack.org/#/c/95987/

The upstream repository is an extraction of the middleware from the
respective repositories maintaining the commit history.

Change-Id: I44a0d59999c720fe9eee99e322138818080b41e0