Add pre-requisites needed for diskimage-builder
For nodepool to use diskimage-builder, it needs a set of pre-reqs
on the nodepool host. Install them.

Also create /var/lib/nodepool for building images.  Testing shows test
images were coming in at about 2.6GB each. If we figure two images for
precise and trusty and one each for centos6 and f20 that is about
15GB. Using the ephemeral partition seems perfect for this as we can
always rebuild an image so they don't have to stick around say after a
nodepool host rebuild.

Change-Id: I725cad999f795a556fd89818578b78b9114fb110