List projects not restricted to pypi.openstack.org
Currently, projects listed in projects.txt are not allowed to use
pip to download requirements from pypi.python.org, and are limited
only to pypi.openstack.org. This patch adds an exception list
where stackforge projects can be listed that will be allowed to
fall back to using pypi.python.org for any requirements not found
in pypi.openstack.org.

Change-Id: Ic69b71e2485108a893a395073eb049e828ad279f