Add proposed/* branches support to merge_tags.sh
Add support for pre-release branches named "proposed/foo" instead of
just "milestone-proposed". Note that only one *proposed* branch should
be present at once in the repository, script won't merge tags if it
finfs more than one.

Also ensures that we exit without a noisy error in the case there is
no *proposed* branch (for example when the script is called on a stable
branch release).

Change-Id: Idbf3b25d1575de1698ed8cfdc1b74e3484391f27