Add proposed/* branches support to merge_tags.sh
Rather than trying to merge tags only if the tag appears to be pushed on
milestone-proposed, merge tags only if the tag appears to not be pushed
on a stable branch. This allows to support pre-release branches named
"proposed/foo" instead of just "milestone-proposed".

Also avoids exiting with a noisy "malformed object name" error in the
case there was no milestone-proposed branch (which was the case when the
script is called on a stable branch release).

Change-Id: Idbf3b25d1575de1698ed8cfdc1b74e3484391f27