Add Monasca sub projects
These sub projects together with python-monascaclient comprise the Monasca monitoring project.

The Monasca name is new, so the projects are being renamed from names like mon-notification to monasca-notification.

Signing the Open Stack Contributer Agreement is required to contribute to Monasca.

The Monasca chef cookbook projects are being put in the group monasca and not openstack-chef because Monasca is not an official OpenStack project.

Some sub projects don't have builds because their builds won't work in Stack Forge. We wanted to get all of the sub projects in and then fix the builds later.

I have created all of the project on pypi that are needed.

Change-Id: Iea2d47b0878439c580b9607fa50b4ea3b98929f8