Run bandersnatch mirror on the mirror host
Bandersnatch is more efficient than pypi-mirror. Instead of using
jenkins slaves to do this, just run it on a 5 minute cron.

Overlay the location of the files on the current directory, so that
http://pypi.openstack.org/openstack and http://pypi.openstack.org/simple
will both work in parallel to ease transition. This will require a
manually move of a dir on the filesystem.

Change-Id: If5935b356e222c2f4d474a2cec8add3cc66b6010