Add two more useful ansible playbooks
One of the more complex activities we have to perform is renaming
projects. Specifically, this involves several steps of deleting
and moving directories all over the place. The rename.yaml playbook
handles this:

  ansible-playbook /etc/ansible/rename.yaml \
    --extra-vars='project=stackforge/designate newname=openstack/designate'

Also, when we get rid of hosts, we need to clean them from puppet's cert
list and also from puppetdb's list of nodes:

  ansible-playbook /etc/ansible/delete_from_puppet.yaml \
    --extra-vars='node=elasticsearch01.openstack.org'

Change-Id: Idc5d4f849b7b08d7df6a6807f1cb83a22d33f43a