Gate against pecan template scaffolding tests.
Pecan provides a "quickstart" command, `pecan create`, which can be used to
build out a new pecan application from a template.  These tests build a new
project, start an HTTP server, and make assertions about its responses to
ensure that changes to pecan have not broken the application code in the
template.

Change-Id: I60337c47b9e21eefaad43327cfaa01dea8eef72e