Add cross-project unit tests for oslo libraries
Add non-voting jobs to run the unit tests of a consuming project with
the unreleased master version of the oslo library. The new jobs are
based on the current dependencies.

Add a tool to help check the test configurations, applying some basic
rules about which sorts of projects gate on which other projects.

Clean up some of the old jobs that were doing similar testing in a
different way and replace them with the new jobs for consistency.

Change-Id: If56a772950434a85eaf323ba320f8e84c45744c8