Align prepare_node.sh with prepare-node in dib
In the dib branch, we made some improvements to the prepare-node
script that are useful to backport in. The exit codes processing
gets us more accurate failures when puppet is running due to stupid
puppet error codes. The install_resolv_conf lets us run the code
in things like docker and dib.

Change-Id: I4720a7ebe71a7d2097d25b18aef62ae5d3238477