Setting environment variables to ensure we execute basic tripleo overcloud upgrade tests.
Tripleo is adding tests of overcloud upgrades to gate.
The execution of these tests are controlled by the value of OVERCLOUD_UPGRADE_TEST.
Setting this env var in the overcloud jobs so this test will be executed.
It should be noted that the upgrade test adds approx 25 minutes to the jobs.
Can set to 0 / create variants without this value to have non-upgrade tests.

Change-Id: Icf263bd68d37b82e7f770edea97bfbbf4fc0a1af