Publish specs jobs for *-specs
The gate jobs already publish the specs as drafts and it
would be very beneficial to actually see those online at a permanent
place - rather than going to github.

This patch:
* Sets up a new site specs.openstack.org
* Creates new job-template '{name}-publish-specs' for publishing specs
* Creates new specs-job job-group for gating and publishing specs
* Converts all *-specs repositories to use the new job-group
* Creates Zuul specs-job macro to use these jobs
* Replaces those *-specs repositories that use already docs and python27
  gates to use the new job-group and thus get published.

The patch does not take care of publishing of repositories that do not
use both the docs and python27 gates, this will be done separately.

implements bp publish-specs

Co-Authored-By: Andreas Jaeger <aj@suse.de>
Change-Id: Icc1c780cea8cd24c7a108429f84b2450056c6eed