Added NodeJS and NPM to jenkins node
It turns out that trusty has a recent-enough version of node
and NPM to allow us to remove the nodeenv dependency in
our javascript builds. This is the first step in doing this, to
get the core javascript build tools onto our jenkins nodes.

Note that a symlink for node is also created, because debian
has a different module called "node" for which there's a
namespace conflict. Since it's unlikely that the node package
(Amateur Packet Radio Node program) will ever be used by
openstack I feel that this is safe to do, especially since
some npm packages (such as phantomjs) unnecessarily
hardcode the name of the node executable.

Change-Id: I2d4321a67b51cc08db8bbc1d9ba07d27c4c7906e