Fix propose_translation runs
The propose_translation runs use "#!/bin/bash -xe" - thus
exit when an error occurs.

It's ok for the grep in filter_commit to not found anything, thus
ignore exit results. Otherwise the script will exit with error code
which is not the intented behaviour.

This unbreaks all propose_translation jobs. My tests were done without
"set -e" in the environment ;(

Change-Id: I8fff0bc42b45b356dd1ca3ab9b2586628ff59a57