Add Nova - oVirt Driver project to StackForge
Nova - oVirt Driver will make use of a Cluster in oVirt Data Center as a Compute in Nova.

Nova compute proxy connects to oVirt and exposes configured clusters as compute hosts. This allows

* a single oVirt driver to model multiple Clusters in oVirt as multiple nova-compute nodes.
* the oVirt driver to be configured to represent a set of clusters as compute nodes.

The glance images are imported as templates into oVirt, which are then used to create VMs in oVirt.

Change-Id: I9fcdc37c4bd3e205ac8fad334bebac35c83b2dfc