Enable a nonvoting puppet3 test.sh for infra/config
I'm not entirely sure this is even a good idea. But it should let us
see what is failing for puppet3. This basically reverts the change
that jeblair did to fix the puppet-apply test when it was running
in puppet 3 mode. Those outputs, while verbose, were useful for finding
where the errors in puppet 3 are.

I don't particularly like the logic for 'puppet 3' in test.sh. I could
see putting it in a builder. I didn't want to replicate facter in jjb.

Change-Id: Ib0af21eb4d78c264b574d73928698eb0292ad424