Add experimental py26 requirements integration job
An experimental py26 requirements integration job will help us
verify that packages are installable in a python26 environment.

This is far from working -- we've never tried devstack on centos6,
there may be a lot to do, or it may not be possible.  But this
will let us check the current situation.

Change-Id: If1f980447964fc56abc2abe7abb20ffd4d40a93d