Set up stunnel on the jenkins masters
This sets up the stunnel service in preparation for the switch to SSL
for gearman. The jenkins gearman plugin doesn't support connecting to
a gearman server over SSL, so instead it will connect locally to the
listening stunnel service which will deal with the SSL wrapping of
the TCP connection.

The service is currently stopped, as it isn't needed until we hit
the big SSL switch.

When ready a patch set is required to start the stunnel service:
    ensure => running

And we need to be sure the SSL cert, key _and_ CA is placed
in the required locations (2 of the three should already be in
place for apache):
    $prv_ssl_cert_file
    $prv_ssl_key_file
    $ssl_chain_file

Change-Id: I88cabfb79003be812215125141dd83122db0c5f7