Set up stunnel on the jenkins masters
This sets up the stunnel service in preparation for the switch to SSL
for gearman. The jenkins gearman plugin doesn't support connecting to
a gearman server over SSL, so instead it will connect locally to the
listening stunnel service which will deal with the SSL wrapping of
the TCP connection.

This is implemented using the puppetlabs::stunnel module, but using
an openstack_project::stunnel module as a shim to simplify for our
usage.

Gearman in SSL mode requires a client identify itself with an SSL
cert, key and CA file. We need to make sure these exist before
going live.

Change-Id: I88cabfb79003be812215125141dd83122db0c5f7