Create a repository for additional Tuskar-UI plugins
For now this is just a plugin for Tuskar-UI that adds integration with
Satelitte 6 to it.  It's separate from the rest of Tuskar-UI for two
reasons, one is to show off and develop the plugin capabilities of
Tuskar-UI, the second is that it is of limited interest for users
outside of Red Hat. We expect to have more of such small plugins, so I
renamed the repository to "extras".

Change-Id: I1efb770a10c2fe940749369e67f127d73e69243d