Create new repository for Tuskar-UI Satelitte 6 plugin
Create a new repository for the Tuskar-UI Satelitte 6 project.

This is a plugin for Tuskar-UI that adds integration with Satelitte 6
to it.  It's separate from the rest of Tuskar-UI for two reasons, one
is to show off and develop the plugin capabilities of Tuskar-UI, the
second is that it is of limited interest for users outside of Red Hat.

Change-Id: I1efb770a10c2fe940749369e67f127d73e69243d