Add firewall traffic on port 27410 to 27510.
This change is specifically required for tripleo CI, see the change ID
listed below for more details.

The allows seed VM to attempt to send a Heat signal to a mocked Heat
service on the Jenkins slave where devtest_*.sh runs.

We open up a port to accept that incoming traffic.

Required for: Ibfd20bf632cdf3381615dad68ecf620685553d5f

Change-Id: I9e85f34339d7943a3a37a3b6772b9eb8ac85bf6c