Add git-upstream as new project to stackforge
This patch adds the config needed to create and maintain git-upstream
in stackforge.

git-upstream is an open source Python application that can be used to
keep in sync with upstream open source projects, mainly OpenStack.

git-upstream provides new git subcommands to support rebasing of
local-carried patches on top of upstream repositories. It provides
commands to ease the use of git for who needs to integrate big
upstream projects in their environment.
The operations are performed using Git commands.

See also:
 * https://github.com/dguerri/git-upstream
 * https://pypi.python.org/pypi/git-upstream

Change-Id: I16f71cbe7178b9e98bed2ee0a9c905b5ee9bc395