Update version of puppetlabs-stdlib in install_modules.sh
Without this patch, install_modules.sh will attempt to install
puppetlabs-stdlib 3.2.0 and fail because puppetlabs-concat installs
puppetlabs-stdlib > 4. This is a problem because it causes inconsistency
between the version listed in install_modules.sh and the actual installed
version, and it causes errors during subsequent runs of the shell script. This
patch resolves the problem by updating the version listed in
install_modules.sh to the latest version.

Closes-Bug: 1324645
Change-Id: I7822597785946330346c476753ff60ae1b345014