Allow comments after "recheck no bug"
"recheck no bug" is not to be encouraged, but there are often
circumstances where it makes the most sense.  Particularly if you've
been away from a change for some time and jenkins has attempted to
verify.  Digging into the reason for a failure perhaps days after the
test is really just a waste of time given how fast things move.

It would be nice to be able to put some sort of comment showing you
have some sane reason for doing this.  See also [1]

[1] http://lists.openstack.org/pipermail/openstack-dev/2014-July/040416.html

Change-Id: If79ff5b7f7fde3e1a92f650e12678a9cea4256da