namos - openstack infra device and service manager
The stackforge project to implment a new service
called namos for managing openstack infra devices,
and it's association with services components.
It provides the complete control for admins to add,
maintain and remove a infra devices such as
storage arrays, hypervisors, network switches in
a given cloud setup.
As well to manange the association between these
devices and openstack service components.

An admin portal will be provided in the horizon under
admin group.

Change-Id: I0f649a7909998d8b0770d738ff659438e3d2918a