namos - openstack infra device and service manager
An openstack service provided for the admin to manage
openstack infra devices, services and it's components.
It provides the complete control to the admin to add,
maintain and remove a infra device say storage arrays,
hypervisors, network switches to the openstack as well
to manange the association between these devices and
openstack services.

An admin portal will be provided in the horizon under
admin group.

Change-Id: I0f649a7909998d8b0770d738ff659438e3d2918a