Refactor the various artifact upload scripts into one
There are different scripts to upload jenkins plugins
and java artifacts.

The intent of this change is to have a common script
for uploading the above artifacts for better code
reusability, and parameterize things like artifact
repositories URLs and credentials for a better
downstream consumption.
The modifications to the maven and jenkins JJB
definitions are simply to show how this change
would look like in practice.

Change-Id: I222e76223f7e5ddf01e6675e13de3391c2914c6c