Refactor the various artifact upload scripts into one and support parameterization of values
There are different scripts to upload different types of
artifacts (jenkinsci-upload.sh for jenkins plugins,
mavecentral-upload.sh for java packages,
pypi-tarball-upload.sh for pypi tarballs and
pypi-wheel-upload.sh for wheels).

The intent of this change is to have a common script
for uploading the above artifacts for better code
reusability, and parameterize things like artifact
repositories URLs and credentials for a better
downstream consumption.

Change-Id: I222e76223f7e5ddf01e6675e13de3391c2914c6c