Publish additional specs repositories
Add publish job for each specs repository that already gates the docs
and does not use the python27 jobs (and thus cannot use the template
specs-jobs). This is done only for "official" OpenStack repositories but
not for stackforge ones.

Also, add python27 job for glance-specs and use the template. 'tox -e
python27' is setup properly and passes.

For cinder-specs, use the specs-jobs template in both projects.yaml and
layout.yaml.

Change-Id: Iba0badd60a70e467e076c8c829c1e8d0dd6b1591