Add logging to manage-projects calls
Sometimes there is a need to debug why a project is not being
created or ACLs are not properly refreshed.
The way we call manage-projects leaves no trace of debug
messages to inspect failures.

Add a -v flag to the manage-projects calls to display
verbose output, and send logging to syslog and to a
custom logfile

Change-Id: I389bd2b128a82a745f6946cbe87c22aa208c161c