Sync files from openstack-manuals with other repos
This is similar to the propose-requirements-updates job: Run the job
propose-openstack-manuals-updates as post job and copy files from
openstack-manuals via script to some other repositories.

The files to be copied are right now the glossary files and they are
copied to operations-guide and security-doc. The copying of the glossary
allows to use openstack-manuals as master reference and also sharing of
translations.

implements bp common-glossary-setup

Change-Id: I3f2d6238c17f45dd149ec282270227a14c7dd1b1