Use a current Trusty image name in hpcloud
HPCloud "deprecates" images and renames them. This appears to also
change the image uuid in the process. Unfortunately Ubuntu has not
updated
https://cloud-images.ubuntu.com/releases/streams/v1/com.ubuntu.cloud:released:hpcloud.json
so I cannot currently verify the current image or the deprecated image
are correct (at least not fully, I can in theory check the hash of the
deprecated image and assume that if the hashes match we are fine even if
the uuids are different).

Go with the new 14.04.1 image which is current and not deprecated. We
problably want to confirm with ubuntu before we merge this change
though.

Change-Id: I979f1b90fa5dd4e5889b58539208303b78feb115