Use a current Trusty image name in hpcloud
HPCloud "deprecates" images and renames them. This appears to also
change the image uuid in the process. I have checked the image uuid
against
https://cloud-images.ubuntu.com/releases/streams/v1/com.ubuntu.cloud:released:hpcloud.json
and according to that file's gpg signature the uuid is correct.
Note I don't have a chain of trust to the signature used so can only
verify the signature not the signer. I also looked for checksums in
the signed json file but they don't seem to exist so I can only check
the uuid.

Go with the new 14.04.1 image which is current and not deprecated. We
problably want to confirm with ubuntu before we merge this change
though.

Change-Id: I979f1b90fa5dd4e5889b58539208303b78feb115