Add ansible node launching
Adding a server is currently a manual task requiring root access. We'd
much prefer for that to be able to be driven from commits in git like
everything else. Add some ansible library modules that can handle what
we want from launch_node (have been submitted to ansible upstream) as
well as a set of playbooks that can be run to create and ensure the
existence of the servers we have in our infrastructure.

Change-Id: Ica6f738b066ed834d554c87db0c29da1ad2c21b3