Add ansible node launching - don't merge
Adding a server is currently a manual task requiring root access. We'd
much prefer for that to be able to be driven from commits in git like
everything else. Add some ansible library modules that can handle what
we want from launch_node (have been submitted to ansible upstream, see:
https://github.com/ansible/ansible/pull/8412) as
well as a set of playbooks that can be run to create and ensure the
existence of the servers we have in our infrastructure.

To try this out:

 pip install shade shade-ansible
 git clone git://github.com/ansible/ansible.git --recursive
 (cd ansible/hacking ; source env-setup)
 cd system-config
 ansible-playbook -i testhosts montytest.yaml

Change-Id: Ica6f738b066ed834d554c87db0c29da1ad2c21b3