Consolidate and rename Ironic jobs
This gets rid of the redundant Ironic jobs and configures all Ironic devstack
jobs to run with VIRT_DRIVER=ironic.  This also configures deploy driver and
exposes that in the job name.  To avoid copy/paste of the huge Ironic devstack
shell builder, a temporary macro s added that will be removed once we can
minimally define the Ironic devstack jobs.

Change-Id: I3c984aff36a9663b33db13fc21df48a35ea6264c