Add f20-devstack-virt-preview node type
Create a new node type that is Fedora 20 with the virt-preview
repo enabled.  It will also have packages needed by devstack cached on
the image, including the ones from the virt-preview repo.

The virt-preview repo contains bleeding edge versions of libvirt and
qemu, which is important for testing some nova features that are
implemented around the same time as the underlying support is added to
libvirt and qemu.

http://fedoraproject.org/wiki/Virtualization_Preview_Repository

A dependent commit will create a job that uses this new node type.

Change-Id: Idb445fb5ff302cc428470fbad9cfeb686558f5bf