Don't run large-ops on glance,swift or keystone
large-ops today only directly tests nova, and to the best of my
knowledge we have never found cases where glance keystone or swift broke
us in a performance regression.

This potentially opens up a whole in the gate where one of these
projects can break large-ops, in which case we can fix the issue and add
large-ops to that project.

Since we have left the ideal world where we have infinite nodes, drop
the jobs that have the least value.

Disable job on:
  glance
  glance.store
  swift
  python-glanceclient
  python-keystoneclient
  python-swiftclient
  keystone

Change-Id: Iba5fb982cfa18ce4155bdc0579bb48ee738d06a4