Remove unused tempest-dsvm-neutron-isolated-parallel job
The tempest-dsvm-neutron-isolated-parallel job was added to the
experimental neutron queue to assess whether neutron was ready
to successfully pass parallel tests.

In the meanwhile this job has been removed from the experimental
queue and the neutron full job has been added to the neutron
check and gate queues and set to voting.

The tempest-dsvm-neutron-isolated-parallel can therefore be
safely removed.

Change-Id: Ia2fa88affbfcff28ec23b41452cc2c2feacab33a