namos - Device and service manager
The stackforge project to implement a new service
called namos for managing devices and it's
association with OpenStack services. It provides
the complete control for admins to add, maintain
and remove a cloud datacenter devices such as
storage arrays, hypervisors, network switches in
a given cloud setup. As well to manage the
association between these devices and OpenStack
services.

An admin portal will be provided in the horizon
under admin group.

Change-Id: I0f649a7909998d8b0770d738ff659438e3d2918a