OpenStack Containers Service
Stackforge project for development of a new official OpenStack service under
the existing Compute program for using advanced features of containers in
combination with Nova. This project is starting in Stackforge by the request
of the Nova core team. It will be moved to the openstack/ namespace once it
has been developed to the point of minimum viability.

Change-Id: I661b8aa1cc0d75b2804815045d5d776b70e5489f