puppet-jobs: Pin puppet-unit-2.7 jobs to precise
We still have strange errors with puppet 2.7.x and ruby 1.9.x on trusty
slave. Puppet 2.7 unit are planned to be deprecated on K and removed on
I. This temporary patch is a bit dirty, but this fix the gate for
puppet jobs by pinning 2.7 unit to precise workers.

(Tested locally, see: http://paste.debian.net/plain/117660 )

Change-Id: I5df90ecd55b9a583268039e03ffc6b6361f08de8