Fix Berkshelf 3.x GeCode 3.x dependency
Berkshelf 3.x requires GeCode 3.x.  Default is to install
wrapper gem dep-selector-libgecode which will compile the 3.x
GeCode libraries.  This is a very large project to compile and
will fail at times when memory or cpus are limited.

To work around this, there is a dep-selector-libgecode flag,
USE_SYSTEM_GECODE that will allow the gem to just use the GeCode
system libraries.  The 3.x version of the GeCode libraries are default
on precise.  The 4.x version of GeCode on trusty is not compatible
with Berkshelf 3.x.

* Use precise for cookbook gates until Berkshelf/Chef DK support trusty
* Install the precise GeCode 3.x system libraries when using
Berkshelf 3.x.

Change-Id: I19d16e06e72aff090771c46c15b2b8de164988c8
Closes-Bug: #1357439