Add bare-f20 nodes
We have some testing it makes sense to bring up on f20 nodes, as the
package support is better at this point [1].  Also, building nodes
stops bit-rot in the setup scripts.

I have tested this by hand on a rax f20 node.  There is one dependent
change to use a later version of the postgres puppet modules

[1] https://review.openstack.org/#/c/115752/

Change-Id: Ia1cfffbffd61d35c8db69f592c99dc2cc865d2c3