Update git puppet modules properly.
When using git sourced puppet modules we need to specify the desired ref
as a tag, sha1, or remote branch because install_modules.sh does not
reset local branches to match remote branches.

Do a git remote update instaed of a git fetch to make sure that all
remotes are updated.

Change-Id: I1785af6fd02e6cda8f5f62b54bdb83efb47b7197