Updating node deff for puppet 3 master
A lot of changes have gone on in the last
week or two. Once this lands, a rooter
can spin up a puppet 3 master node and attempt
to run util/puppetmaster_bootstraph.sh

That should configure a lot of the Puppet
system. Next it is required to rm -fr
/var/lib/puppet/ssl.

Then copy over the puppet certificate
files from ci-puppetmaster:

/var/lib/puppet/ssl/certs/ca.pem
/var/lib/puppet/ssl/certs/puppetmaster.o.o.pem
/var/lib/puppet/ssl/private_keys/puppetmaster.o.o.pem

With those files in place we should be able to restart
apache and it should turn on, though it won't work just
yet.

Next we need to land a change to turn puppetdb => true
so that this puppetmaster works with PuppetDB.

We should probably be running puppet initially against
itself.

Change-Id: I37879153c68011c752919675a2d3416383aaf8e5