Set download_cache in pip.conf
In the case that for some reason we end up installing the same package
twice, don't pull it from the internet use a download_cache.

devstack already uses /var/cache/pip as the cache when it pip installs
things. But devstack doesn't do this when tempest is run in grenade (as
tempest is run via tox and PIP_DOWNLOAD_CACHE isn't exported.

Instead of custom wiring up the devstack case do this for all jobs.

Note: In the past there was an issue with local pip cached files being
used in preference over downloading new ones even if the locally cached
version was outside the version spec being requested. This issue needs
to be confirmed as resolved before this patch can be merged.

Change-Id: I97671b327e94cdf5a4a154298327a272776507a2