Add a gate job for zaqar-redis
A new redis drier was recently added to Zaqar. The team would like to
have a test job that checks *all* patches against this driver as well.

This patch adds such job to the infrastructure configs. It's based on
the already existing job but it enables redis by setting `ZAQAR_BACKEND`
in the gate environment.

Change-Id: Icb260f27e01d024fe98cf03f68a7285ec4bb3b71