Allow usage of customized venvs for docs building
Certain projects are using customized virtualenvs for docs
building so it seems appropriate to allow those virtualenvs
to be used (instead of the implicit 'venv') when the project
feels it is appropriate to use it.

To retain the current mechanism the usage of this new option
defaults to the current 'venv' when not specified, when it is
specified the alternate venv is used to build the docs.

Change-Id: I5eb063d50d6f52246981391bca9c51867eb7af50