Trim jenkins::slave package installations further
We are installing packages in jenkins::slave that we do not need on all
of our slaves. Trim the package install down there and move things to
openstack_project::thick_slave where they belong. In particular
zookeeper and php mcrypt packages.

As part of this cleanup remove unused parameter values from
jenkins::params.

Change-Id: I57bf1c61287f74418bb06502f06bcd2989161243