Run the precise nonha job on tripleo projects
Add check-tripleo-novabm-overcloud-precise-nonha back running on tripleo
projects, it was removed some time ago due to capacity problems, we now
have the ability to run this job so should do so. Initially adding back
as non voting as this job isn't as stable as it should be (it suffered
by not getting as much attention by tripleo), will switch back to voting
once problems have been resolved.

This will also make it non voting for the other projects it runs
against, it probably should be anyways until its more reliable.

Change-Id: I9900626afdc342ad9522f124475fbd611faec3b4